PR: `bugfix/fix_1_5` → `feature/multi-gpu`

Check PR summary and code review for more infos.

+++ src/nb/experiments/perplexity_eval_RM_specialDeberta.py
@@ -156,7 +156,7 @@ def main() -> None:
     tokenizer = AutoTokenizer.from_pretrained(args.model)
     device_map = "auto" if args.device in ("auto", "cuda") else str(device)
     model = AutoModelForSequenceClassification.from_pretrained(
-        args.model, torch_dtype=dtype, device_map=device_map
+        args.model, dtype=dtype, device_map=device_map
     )
     # .to() intentionally omitted: device_map handles placement.
     model.eval()
